refactor: split module
