+++ modules/config.ts
@@ -1,0 +1,3 @@
+export const args = Deno.args;
+export const [cmd] = args;
+export const cwdSettingFile = `${Deno.cwd()}/deno-scripts.json`;
